requirements.txt
baostock==0.8.8
numpy==1.21.6
openpyxl==3.0.10
pandas==1.3.5
PyYAML==6.0
TA_Lib==0.4.25
